split tkgui.py into multiple files
