+++ .gitignore
@@ -3,6 +3,8 @@ venv*/
 .idea/
 settings.ini
 
+.DS_Store
+
 ### venv template
 # Virtualenv
 # http://iamzed.com/2009/05/07/a-primer-on-virtualenv/
